experiments/models/projection01/requirements.txt
numpy
pandas
scikit-learn
matplotlib
umap-learn
scipy
